really fixed the issue where pngtochr was swapping bits.
The problem was with the get_pixel() and planar bit math.
On 8bpp mode the get_pixel() was correct, less than 8 bits and it would reverse the bits.
The planar math was reversing the bits which canceled out the error for sub-8bpp modes, but created mirroring on 8bpp modes.